requirements.txt
easyocr==1.7.2
PyMuPDF==1.24.13
configparser==7.1.0
pyinstaller==6.11.0
python-docx==1.1.2
